+++ test/modules/datatype/alpha-numeric.test.ts
@@ -34,10 +34,60 @@ describe("datatype.alphaNumeric", () => {
     });
   });
 
+  describe("prefix argument", () => {
+    it("prefix = undefined. should return an alphanumeric string", () => {
+      const value = modules.datatype.alphaNumeric({
+        prefix: undefined,
+        length: 10,
+      });
+
+      expect(value).toHaveLength(10);
+    });
+
+    it("prefix = 'foo'. should return an alphanumeric that starts with 'foo'", () => {
+      const value = modules.datatype.alphaNumeric({
+        prefix: "foo",
+        length: 10,
+      });
+
+      expect(value.startsWith("foo")).toBe(true);
+      expect(value).toHaveLength(10);
+    });
+  });
+
+  describe("case argument", () => {
+    it("case = undefined. should return an mixed case alphanumeric string", () => {
+      const value = modules.datatype.alphaNumeric({
+        case: undefined,
+        length: 10,
+      });
+
+      expect(value).toHaveLength(10);
+    });
+
+    it("case = 'lower'. should return an lower case alphanumeric string", () => {
+      const value = modules.datatype.alphaNumeric({
+        case: "lower",
+      });
+
+      expect(value.toLowerCase()).toBe(value);
+    });
+
+    it("case = 'upper'. should return an upper case alphanumeric string", () => {
+      const value = modules.datatype.alphaNumeric({
+        case: "upper",
+      });
+
+      expect(value.toUpperCase()).toBe(value);
+    });
+  });
+
   describe("banned argument", () => {
     it("banned = 'b'. should return an string without b", () => {
       const val = modules.datatype.alphaNumeric({ length: 10, banned: "b" });
+
       expect(validate(val, ["b"])).toBe(true);
+      expect(val).toHaveLength(10);
     });
 
     it("abnned = ['b', 'a', 'c', 'd', 'e']. should return an string without b, a, c, d and e", () => {